Add customized nona-mask binary
The original from Panotools::Script does not forward all of our arguments properly.

We have provided a patch for this, but it has not yet been released.
This binary file contains that patch.

https://sourceforge.net/p/panotools/Panotools-Script/merge-requests/1/

A new `Binaries` module has been added so we can reference the abstraction cleanly and change it back to just `nona-mask` globally later on.

Signed-off-by: stowersjoshua <stowers.joshua@live.com>